+++ src/adaptors/zapper.ts
@@ -11,10 +11,10 @@ interface ZapperResult {
   zp: number;
 }
 
-export default async function geZapperData(address: string) {
+export default async function getZapperData(address: string) {
   const jsonData = (await fetcher(
     'GET',
-    `https://api.zapper.fi/v1/gamification/users/${address.toLowerCase()}?api_key=562eee97-e90e-42ac-8e7b-363cdff5cdaa`
+    `https://api.zapper.fi/v1/gamification/users/${address.toLowerCase()}?api_key=cbaadb5b-92d2-4479-9a96-ee804989e27a`
   )) as ZapperResult;
 
   return {